refactor(formatter_graphql): fix up code and comments (#23488)
Finally fixed up Fable's work.
