+++ .gitattributes
@@ -2,6 +2,8 @@ crates/oxc_formatter/tests/fixtures/** text=auto eol=lf
 crates/oxc_formatter/tests/fixtures/js/crlf/** text=auto eol=crlf
 crates/oxc_formatter_json/tests/fixtures/** text=auto eol=lf
 crates/oxc_formatter_json/tests/fixtures/jsonc/crlf/** text=auto eol=crlf
+crates/oxc_formatter_graphql/tests/fixtures/** text=auto eol=lf
+crates/oxc_formatter_graphql/tests/fixtures/graphql/crlf/** text=auto eol=crlf
 apps/oxfmt/test/**/fixtures/** text=auto eol=lf
 apps/oxlint/test/fixtures/** text=auto eol=lf
