+++ crates/oxc_formatter_graphql/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "oxc_formatter_graphql"
-version = "0.54.0"
+version = "0.56.0"
 authors.workspace = true
 categories.workspace = true
 edition.workspace = true
