+++ crates/oxc_formatter_graphql/src/print/mod.rs
@@ -1,4 +1,5 @@
-use apollo_parser::{SyntaxKind, SyntaxNode, cst, cst::CstNode};
+use apollo_parser::{cst, cst::CstNode};
+
 use oxc_formatter_core::{
     Buffer, Format, Formatter,
     builders::{FormatWith, empty_line, hard_line_break, if_group_fits_on_line, soft_line_break},
@@ -8,8 +9,9 @@ use oxc_span::Span;
 
 use crate::{
     comments::{
-        flush_leading_comments, flush_trailing_inside_comments, is_suppressed_before,
-        write_dangling_comments, write_suppressed_node, write_trailing_same_line_comment,
+        Gap, classify_gap, flush_leading_comments, flush_trailing_inside_comments,
+        is_suppressed_before, write_dangling_comments, write_suppressed_node,
+        write_trailing_same_line_comment,
     },
     context::GraphqlFormatContext,
 };
@@ -17,7 +19,11 @@ use crate::{
 pub mod common;
 pub mod definition;
 pub mod selection;
+mod sig;
+pub mod string;
 pub mod value;
+
+use sig::{sig_end, sig_start};
 
 pub type GraphqlFormatter<'buf, 'a> = Formatter<'buf, 'a, GraphqlFormatContext<'a>>;
 
@@ -40,68 +46,6 @@ where
     T: Fn(&mut GraphqlFormatter<'_, 'a>),
 {
     FormatWith::new(formatter)
-}
-
-/// Whether `kind` is a trivia token. In GraphQL, commas are trivia too.
-pub fn is_trivia(kind: SyntaxKind) -> bool {
-    matches!(kind, SyntaxKind::WHITESPACE | SyntaxKind::COMMENT | SyntaxKind::COMMA)
-}
-
-/// Start offset of the first significant (non-trivia) token within `node`.
-///
-/// apollo-parser attaches pending trivia to the node that is open when the next
-/// significant token is consumed, so `node.text_range().start()` may point at a
-/// comment that logically precedes the node. All layout decisions use significant
-/// token positions instead.
-pub fn sig_start(node: &SyntaxNode) -> u32 {
-    let node_end = u32::from(node.text_range().end());
-    let mut tok = node.first_token();
-    while let Some(t) = tok {
-        if u32::from(t.text_range().start()) >= node_end {
-            break;
-        }
-        if !is_trivia(t.kind()) {
-            return t.text_range().start().into();
-        }
-        tok = t.next_token();
-    }
-    node.text_range().start().into()
-}
-
-/// End offset of the last significant (non-trivia) token within `node`.
-pub fn sig_end(node: &SyntaxNode) -> u32 {
-    let node_start = u32::from(node.text_range().start());
-    let mut tok = node.last_token();
-    while let Some(t) = tok {
-        if u32::from(t.text_range().end()) <= node_start {
-            break;
-        }
-        if !is_trivia(t.kind()) {
-            return t.text_range().end().into();
-        }
-        tok = t.prev_token();
-    }
-    node.text_range().end().into()
-}
-
-/// Significant-token span of `node`.
-pub fn sig_span(node: &SyntaxNode) -> Span {
-    Span::new(sig_start(node), sig_end(node))
-}
-
-/// Start offset of a closing delimiter: the token's start when present,
-/// the container's significant end otherwise (error-resilient fallback).
-pub fn closing_token_start(
-    token: Option<apollo_parser::SyntaxToken>,
-    container: &SyntaxNode,
-) -> u32 {
-    token.map_or_else(|| sig_end(container), |t| t.text_range().start().into())
-}
-
-/// Source slice of `node`'s significant span, carrying the arena lifetime.
-pub fn node_text<'a>(f: &GraphqlFormatter<'_, 'a>, node: &SyntaxNode) -> &'a str {
-    let span = sig_span(node);
-    f.context().source_text().slice_range(span.start, span.end)
 }
 
 /// How consecutive sequence items are separated.
@@ -146,8 +90,7 @@ where
                 .filter(|c| c.start < start)
                 .map_or(start, |c| c.start);
             let is_blank = preserve_blank
-                && crate::comments::classify_gap(f.context().source_text().bytes_range(pe, anchor))
-                    == crate::comments::Gap::Blank;
+                && classify_gap(f.context().source_text().bytes_range(pe, anchor)) == Gap::Blank;
             match separator {
                 SeparatorKind::Hard => {
                     if is_blank {
