+++ crates/oxc_formatter_graphql/tests/fixtures/graphql/block-string.graphql
