+++ crates/oxc_formatter_graphql/tests/fixtures/graphql/commas.graphql
