+++ crates/oxc_formatter_graphql/tests/fixtures/graphql/crlf/ignore.graphql
@@ -1,0 +1,9 @@
+# prettier-ignore
+type   Weird   {
+  a:   String
+}
+
+# oxfmt-ignore
+query   Q   {
+  weird   {   layout   }
+}
