+++ crates/oxc_formatter_graphql/tests/fixtures/graphql/crlf/options.json
@@ -1,0 +1,4 @@
+[
+  { "endOfLine": "lf" },
+  { "endOfLine": "crlf" }
+]
