+++ crates/oxc_formatter_graphql/tests/fixtures/graphql/empty-values.graphql
