+++ crates/oxc_formatter_graphql/tests/fixtures/graphql/executable-descriptions.graphql
