+++ crates/oxc_formatter_graphql/tests/fixtures/graphql/implements-width.graphql
