+++ crates/oxc_formatter_graphql/tests/fixtures/graphql/string-escapes.graphql
