+++ crates/oxc_formatter_graphql/tests/fixtures/graphql/trailing-comments.graphql
